atualização de testes unitários, adição do Redis

+++ lista-de-tarefas/src/main/java/com/listadetarefas/service/TarefaService.java
@@ -1,46 +1,61 @@
 package com.listadetarefas.service;
 
+import com.listadetarefas.client.UsuarioClient;
+import com.listadetarefas.config.RabbitMQConfig;
 import com.listadetarefas.dto.TarefaCreateRequestDTO;
 import com.listadetarefas.dto.TarefaUpdateRequestDTO;
 import com.listadetarefas.dto.TarefaResponseDTO;
+import com.listadetarefas.dto.UsuarioDTO;
+import com.listadetarefas.event.NotificacaoTarefaEvent;
 import com.listadetarefas.exception.TarefaNaoEncontradaException;
 import com.listadetarefas.model.Prioridade;
 import com.listadetarefas.model.Status;
 import com.listadetarefas.model.Tarefa;
 import com.listadetarefas.repository.TarefaRepository;
 import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
+@Slf4j
 @Service
+@RequiredArgsConstructor
 public class TarefaService {
 
     private final TarefaRepository tarefaRepository;
+    private final UsuarioClient usuarioClient;
+    private final RabbitTemplate rabbitTemplate;
 
-    public TarefaService(TarefaRepository tarefaRepository){
-        this.tarefaRepository = tarefaRepository;
-    }
-
+    @Cacheable(value = "tarefas", key = "#usuarioId", condition = "#status == null && #prioridade == null")
     public List<TarefaResponseDTO> buscarTarefasPorStatusOuPrioridade(Long usuarioId, Status status, Prioridade prioridade) {
 
+
         if (status != null && prioridade != null) {
+            log.info("Buscando tarefas do usuário {} no banco de dados...", usuarioId);
             return tarefaRepository.findByUsuarioIdAndStatusAndPrioridade(usuarioId, status, prioridade)
-                    .stream().map(this::converterParaDTO).toList();
+                    .stream().map(this::converterParaDTO).collect(Collectors.toList());
         }
 
         if (status != null) {
+            log.info("Buscando tarefas do usuário {} no banco de dados...", usuarioId);
             return tarefaRepository.findByUsuarioIdAndStatus(usuarioId, status)
-                    .stream().map(this::converterParaDTO).toList();
+                    .stream().map(this::converterParaDTO).collect(Collectors.toList());
         }
 
         if (prioridade != null) {
+            log.info("Buscando tarefas do usuário {} no banco de dados...", usuarioId);
             return tarefaRepository.findByUsuarioIdAndPrioridade(usuarioId, prioridade)
-                    .stream().map(this::converterParaDTO).toList();
+                    .stream().map(this::converterParaDTO).collect(Collectors.toList());
         }
 
         return tarefaRepository.findByUsuarioId(usuarioId)
-                .stream().map(this::converterParaDTO).toList();
+                .stream().map(this::converterParaDTO).collect(Collectors.toList());
     }
 
     public TarefaResponseDTO buscarPorId(Long id, Long usuarioId) {
@@ -49,43 +64,90 @@ public class TarefaService {
     }
 
     @Transactional
+    @CacheEvict(value = "tarefas", key = "#usuarioId")
     public TarefaResponseDTO criarTarefa(TarefaCreateRequestDTO request, Long usuarioId) {
+        Tarefa novaTarefa = persistirNovaTarefa(request, usuarioId);
 
-        Tarefa novaTarefa = new Tarefa();
-        novaTarefa.setNome(request.nome());
-        novaTarefa.setStatus(Status.NAO_CONCLUIDA);
-        novaTarefa.setPrioridade(request.prioridade());
+        orquestrarNotificacao(usuarioId, "TAREFA_CRIADA", novaTarefa.getNome());
 
-        novaTarefa.setUsuarioId(usuarioId);
-
-        novaTarefa = tarefaRepository.save(novaTarefa);
         return converterParaDTO(novaTarefa);
     }
 
     @Transactional
-    public TarefaResponseDTO atualizarTarefaParcialmente(Long id, Long usuarioId, TarefaUpdateRequestDTO request) {
-        Tarefa tarefa = buscarTarefaPorIdEUsuario(id, usuarioId);
+    @CacheEvict(value = "tarefas", key = "#usuarioId")
+    public TarefaResponseDTO atualizarTarefa(Long id, TarefaUpdateRequestDTO request, Long usuarioId) {
+        Tarefa tarefa = tarefaRepository.findById(id)
+                .orElseThrow(() -> new TarefaNaoEncontradaException(id, usuarioId));
+
+        if (!tarefa.getUsuarioId().equals(usuarioId)) {
+            throw new RuntimeException("Acesso negado a esta tarefa");
+        }
 
         if (request.nome() != null && !request.nome().trim().isEmpty()) {
             tarefa.setNome(request.nome());
         }
-
         if (request.status() != null) {
             tarefa.setStatus(request.status());
         }
-
         if (request.prioridade() != null) {
             tarefa.setPrioridade(request.prioridade());
         }
 
         tarefa = tarefaRepository.save(tarefa);
+
+        orquestrarNotificacao(usuarioId, "TAREFA_ATUALIZADA", tarefa.getNome());
+
         return converterParaDTO(tarefa);
     }
 
     @Transactional
+    @CacheEvict(value = "tarefas", key = "#usuarioId")
     public void deletarTarefa(Long id, Long usuarioId) {
-        Tarefa tarefa = buscarTarefaPorIdEUsuario(id, usuarioId);
+        Tarefa tarefa = tarefaRepository.findById(id)
+                .orElseThrow(() -> new TarefaNaoEncontradaException(id, usuarioId));
+
+        if (!tarefa.getUsuarioId().equals(usuarioId)) {
+            throw new RuntimeException("Acesso negado a esta tarefa");
+        }
+
+        String nomeTarefaDeletada = tarefa.getNome();
+
         tarefaRepository.delete(tarefa);
+
+        orquestrarNotificacao(usuarioId, "TAREFA_DELETADA", nomeTarefaDeletada);
+    }
+
+
+    private Tarefa persistirNovaTarefa(TarefaCreateRequestDTO request, Long usuarioId) {
+        Tarefa tarefa = new Tarefa();
+        tarefa.setNome(request.nome());
+        tarefa.setStatus(Status.NAO_CONCLUIDA);
+        tarefa.setPrioridade(request.prioridade());
+        tarefa.setUsuarioId(usuarioId);
+        return tarefaRepository.save(tarefa);
+    }
+
+    private void orquestrarNotificacao(Long usuarioId, String tipoEvento, String tituloTarefa) {
+        try {
+            UsuarioDTO usuario = usuarioClient.buscarUsuarioPorId(usuarioId);
+
+            NotificacaoTarefaEvent evento = new NotificacaoTarefaEvent(
+                    usuario.email(),
+                    usuario.nome(),
+                    tipoEvento,
+                    tituloTarefa
+            );
+
+            enviarMensagemParaFila(evento);
+
+        } catch (Exception e) {
+            log.error("Erro ao orquestrar a notificação '{}' para a tarefa '{}'. O e-mail não será enviado.", tipoEvento, tituloTarefa, e);
+        }
+    }
+
+    private void enviarMensagemParaFila(NotificacaoTarefaEvent evento) {
+        rabbitTemplate.convertAndSend(RabbitMQConfig.FILA_NOTIFICACOES, evento);
+        log.info("Evento '{}' da tarefa '{}' postado na fila com sucesso.", evento.tipoEvento(), evento.tituloTarefa());
     }
 
     private Tarefa buscarTarefaPorIdEUsuario(Long id, Long usuarioId) {
